% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis.R
\name{bernstein_basis}
\alias{bernstein_basis}
\title{Create a Bernstein basis}
\usage{
bernstein_basis(K)
}
\arguments{
\item{K}{The degree of the Bernstein basis.}
}
\value{
An MTRBasis list for the Bernstein basis.
}
\description{
This function generates a Bernstein basis of degree K.
}
