% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis.R
\name{bernstein_polynomial}
\alias{bernstein_polynomial}
\title{kth Bernstein basis polynomial of degree K}
\usage{
bernstein_polynomial(u, k, K)
}
\arguments{
\item{u}{The point at which to evaluate the polynomial.}

\item{k}{The index of the Bernstein basis polynomial.}

\item{K}{The degree of the Bernstein basis.}
}
\value{
The value of the kth Bernstein basis polynomial of degree K at point u.
}
\description{
This function calculates the kth Bernstein basis polynomial of degree K at a given point u.
}
