% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{compute_average_weights}
\alias{compute_average_weights}
\title{Compute Average Weights for Target Parameters}
\usage{
compute_average_weights(tp, dgp = NULL)
}
\arguments{
\item{tp}{A list containing the name of the target parameter and its integration limits.}

\item{dgp}{An optional DGP object containing the data generating process.}
}
\value{
A data frame containing the average weights.
}
\description{
This function computes the average weights for different target parameters.
}
