% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivlike.R
\name{compute_beta_s}
\alias{compute_beta_s}
\title{Compute Beta_s}
\usage{
compute_beta_s(dgp, slist, param = NA)
}
\arguments{
\item{dgp}{Data Generating Process object.}

\item{slist}{List of s functions.}

\item{param}{Optional parameter for slist.}
}
\value{
A numeric vector of beta_s values.
}
\description{
Compute Beta_s
}
