% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivlike.R
\name{compute_gamma_s}
\alias{compute_gamma_s}
\title{Compute gamma_s for Each Model}
\usage{
compute_gamma_s(bases, dgp, slist, param = NA)
}
\arguments{
\item{bases}{List of basis functions for each model.}

\item{dgp}{Data Generating Process object.}

\item{slist}{List of s functions.}

\item{param}{Optional parameter for slist.}
}
\value{
A list of gamma_s arrays for each model.
}
\description{
Compute gamma_s for Each Model
}
