% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivlike.R
\name{compute_gamma_s_for_basis}
\alias{compute_gamma_s_for_basis}
\title{Compute gamma_s for a given basis}
\usage{
compute_gamma_s_for_basis(basis, d, dgp, slist, param = NA)
}
\arguments{
\item{basis}{Basis functions for a specific model.}

\item{d}{Treatment indicator (0 or 1).}

\item{dgp}{Data Generating Process object.}

\item{slist}{List of s functions.}

\item{param}{Optional parameter for slist.}
}
\value{
A gamma_s array for the specific basis.
}
\description{
Compute gamma_s for a given basis
}
