% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target_parameters.R
\name{compute_single_gamma_star}
\alias{compute_single_gamma_star}
\title{Compute Single Gamma Star}
\usage{
compute_single_gamma_star(tp, basis, d, l, dgp)
}
\arguments{
\item{tp}{Target parameter object.}

\item{basis}{Basis functions for a specific model.}

\item{d}{Treatment indicator (0 or 1).}

\item{l}{Model index.}

\item{dgp}{Data Generating Process object.}
}
\value{
A Gamma_star matrix for the specific basis.
}
\description{
Compute Single Gamma Star
}
