% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target_parameters.R
\name{eval_tp}
\alias{eval_tp}
\title{Evaluate Target Parameter}
\usage{
eval_tp(tp, mtrs, dgp)
}
\arguments{
\item{tp}{Target parameter object.}

\item{mtrs}{List of MTR objects.}

\item{dgp}{Data Generating Process object.}
}
\value{
A numeric value representing the evaluated target parameter.
}
\description{
Evaluate Target Parameter
}
