% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{get_non_zero_weights}
\alias{get_non_zero_weights}
\title{Get Non-Zero Weights from Weight Data Frame}
\usage{
get_non_zero_weights(weight_df, d)
}
\arguments{
\item{weight_df}{Data frame containing weights.}

\item{d}{Treatment indicator (0 or 1).}
}
\value{
A data frame containing non-zero weights.
}
\description{
Get Non-Zero Weights from Weight Data Frame
}
