% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivlike.R
\name{ivslope_indicator}
\alias{ivslope_indicator}
\title{Compute IV Slope Indicator}
\usage{
ivslope_indicator(dgp, support)
}
\arguments{
\item{dgp}{Data Generating Process object.}

\item{support}{Support set for the indicator.}
}
\value{
A list containing the name, s function, support set, and legend titles.
}
\description{
Compute IV Slope Indicator
}
