% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target_parameters.R
\name{late}
\alias{late}
\title{Create LATE Target Parameter}
\usage{
late(dgp, u1, u2, l = 1)
}
\arguments{
\item{dgp}{Data Generating Process object.}

\item{u1}{Lower limit for LATE.}

\item{u2}{Upper limit for LATE.}

\item{l}{Model index (default is 1).}
}
\value{
A TargetParameter object for LATE.
}
\description{
Create LATE Target Parameter
}
