% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtr.R
\name{MTR}
\alias{MTR}
\title{Create an MTR (Marginal Treatment Response) object as a list}
\usage{
MTR(basis, theta)
}
\arguments{
\item{basis}{A list containing basis functions 'a' and 'b'.}

\item{theta}{A matrix of coefficients for the basis functions.}
}
\value{
A list containing the basis functions and their coefficients (theta).
}
\description{
Create an MTR (Marginal Treatment Response) object as a list
}
