% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis.R
\name{MTRBasis}
\alias{MTRBasis}
\title{Create an MTRBasis list}
\usage{
MTRBasis(a, b, ib)
}
\arguments{
\item{a}{A list of functions for the 'a' component.}

\item{b}{A list of functions for the 'b' component.}

\item{ib}{A list of functions for the 'ib' component.}
}
\value{
A list containing the components a, b, and ib.
}
\description{
This function creates an MTRBasis list containing three components: a, b, and ib.
}
