% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivlike.R
\name{olsslope}
\alias{olsslope}
\title{Compute OLS Slope}
\usage{
olsslope(dgp)
}
\arguments{
\item{dgp}{Data Generating Process object.}
}
\value{
A list containing the name and s function.
}
\description{
Compute OLS Slope
}
