% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target_parameters.R
\name{TargetParameter}
\alias{TargetParameter}
\title{Create a Target Parameter Structure}
\usage{
TargetParameter(name, int_limits, int_constant, legendtitle = NULL)
}
\arguments{
\item{name}{Name of the target parameter.}

\item{int_limits}{Function for integration limits.}

\item{int_constant}{Function for integration constant.}

\item{legendtitle}{Optional legend title.}
}
\value{
A list representing the target parameter.
}
\description{
Create a Target Parameter Structure
}
